Add venting routine for oxidizer tank during prelaunch
This was added to engine control, because the oxidizer tank is part of the engine.
Previously, this listened to the ground station for commands.
But it didn't make sense because avionics has control of the vent valve and also
the peripheral to read the tank pressure. The vent control algorithm is also simple,
it's just a check of whether or not the pressure if over a certain precalculated number.